First upload of the NeoStore source code
I'm happy to declare the NeoStore to the Linux world. This is a simple package manager with the goal to replace the Deepin Store, Synaptic and make a united interface for the package managing. After the complete implementation of APT, I'll look in Flatpak, Snap, AppImage and Anbox. Step by step, this will be a famous project and a must of the package managers.